+++ site/controller.php
@@ -2,7 +2,7 @@
 /**
  * @package		Joomla
  * @subpackage	LUPO
- * @copyright	Copyright (C) 2006 - 2014 databauer / Stefan Bauer
+ * @copyright	Copyright (C) databauer / Stefan Bauer
  * @author		Stefan Bauer
  * @link				http://www.ludothekprogramm.ch
  * @license		License GNU General Public License version 2 or later
@@ -22,7 +22,6 @@ jimport('joomla.application.component.helper');
 class LupoController extends JControllerLegacy {
 	function display(){
         $document = JFactory::getDocument();
-
         $app = JFactory::getApplication();
         $params = $app->getParams();
         $uikit = $params->get('lupo_load_uikit_css', "0");
@@ -34,34 +33,48 @@ class LupoController extends JControllerLegacy {
 
 		$view = JRequest::getVar('view');
 		$id = JRequest::getVar('id');
-		
-		if($view=='category'){
-			$model = & $this->getModel();
-			$category = $model->getCategory($id);
-			$games = $model->getGames($id);
-			
-			$view = & $this->getView('Category', 'html');
-			$view->assignRef('category',$category);
-			$view->assignRef('games',$games);
-			
-			$view->display();
-			
-		}elseif($view=='game'){
-			$model = & $this->getModel();
-			$game = $model->getGame($id);
-			
-			$view = & $this->getView('Game', 'html');
-			$view->assignRef('game',$game);
 
-			$view->display();
-			
-		} else {
-			
-			$view = & $this->getView('Categories', 'html');
-			$model = & $this->getModel();
-			$categories = $model->getCategories();
-			$view->assignRef('categories',$categories);
-			$view->display();
+		switch($view){
+			case 'category':
+				$model = &$this->getModel();
+				$category = $model->getCategory($id);
+				$games = $model->getGames($id);
+				$view = &$this->getView('Category', 'html');
+				$view->assignRef('category', $category);
+				$view->assignRef('games', $games);
+				$view->display();
+				break;
+			case 'agecategory':
+				$model = & $this->getModel();
+				$agecategory = $model->getAgecategory($id);
+				$games = $model->getGames($id, 'age_catid');
+				$view = & $this->getView('Agecategory', 'html');
+				$view->assignRef('agecategory',$agecategory);
+				$view->assignRef('games',$games);
+				$view->display();
+				break;
+			case 'game':
+				$model = & $this->getModel();
+				$game = $model->getGame($id);
+				$view = & $this->getView('Game', 'html');
+				$view->assignRef('game',$game);
+				$view->display();
+				break;
+			case 'agecategories':
+				$view = & $this->getView('Agecategories', 'html');
+				$model = & $this->getModel();
+				$agecategories = $model->getAgecategories();
+				$view->assignRef('agecategories',$agecategories);
+				$view->display();
+				break;
+			case 'categories':
+			default:
+				$view = & $this->getView('Categories', 'html');
+				$model = & $this->getModel();
+				$categories = $model->getCategories();
+				$view->assignRef('categories',$categories);
+				$view->display();
+				break;
 		}
 	}
 }