+++ site/views/agecategories/tmpl/default.php
@@ -26,10 +26,10 @@ $menu =JSite::getMenu()->getActive();
 
 <ul class="lupo lupo_categories">
 <?php
-foreach($this->categories as $category){
-	?><li><a href="<?php echo $category['link']?>"><?php echo $category['title']?></a> 
+foreach($this->agecategories as $agecategory){
+	?><li><a href="<?php echo $agecategory['link']?>"><?php echo $agecategory['title']?></a>
 	<?php if($componentParams->get('cats_nbr_games', '1')) { ?>
-		(<?php echo $category['number']?>)</li><?php
+		(<?php echo $agecategory['number']?>)</li><?php
 	}	
 }
 ?>