+++ admin/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 2.0.0 (Januar 27, 2015)
+
+  - listing of agecategories added
+
+### 1.8.0 (Januar 26, 2015)
+
+  - fields added: keywords and genres
+
 ### 1.7.0 (November 19, 2014)
 
   - french translation added